test done for userInformation
